% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-search.R
\name{automatic_error_search}
\alias{automatic_error_search}
\title{An automatic error check and search}
\usage{
automatic_error_search()
}
\value{
Nothing
}
\description{
When passed to `options(error = automatic_error_search`, the function will
run every time R encounters an error. It will check whether you have made
the common mistake of trying to subset a function ('closure'). If it is any
other type of error, it will copy the error message and launch a Google
search.
}
\details{
This function is still being developed and tested and functionality may
change. For example, I may decide to include a list of errors that do not
trigger a Google search.
}
\examples{
\dontrun{
  options(error = automatic_error_search)
}

}
\references{
Adapted from the code published in this Tweet
https://twitter.com/RCoderWeb/status/1358474355492151297
}
