% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize-new-project.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(string, overwrite = FALSE)
}
\arguments{
\item{string}{A character string giving the name of the new folder.}

\item{overwrite}{A boolean equal to TRUE if existing folders and files
should be overwritten. All warnings are supressed. Default is FALSE.}
}
\description{
A wrapper function for creating new directories. Each directory is created
relative to the current working directory.
}
