% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_structure}
\alias{create_dir_structure}
\title{Create the directory structure based on the supplied file list}
\usage{
create_dir_structure(file_list)
}
\arguments{
\item{file_list}{A vector of files}
}
\description{
Create the directory structure based on the supplied file list
}
