% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_all.R
\name{detach_all}
\alias{detach_all}
\title{Detach all non-essential objects from all environments}
\usage{
detach_all()
}
\description{
This function detaches all non-essential objects from all environments and is
a simple rewrite of a function available through the codes provided by
the Choice Modeling Centre in Leeds.
}
\examples{
detach_all()

}
