% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_all_objects.R
\name{detach_all_objects}
\alias{detach_all_objects}
\title{Detach all non-essential objects from all environments}
\usage{
detach_all_objects()
}
\description{
This function detaches all non-essential objects from all environments and is 
a simple rewrite of a function available through the codes provided by
the Choice Modeling Centre in Leeds.
}
