% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{download_beamer}
\alias{download_beamer}
\title{Download the Beamer template from Github}
\usage{
download_beamer(beamer_theme)
}
\arguments{
\item{beamer_theme}{A character string giving the full name of the Github
folder with the theme to be downloaded.}
}
\description{
The function downloads all the files from the Github repository into a
folder with the same name as the subfolder in the online repository
}
