% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{download_from_github}
\alias{download_from_github}
\title{Downloads a folder and its files from Github}
\usage{
download_from_github(directory, user, repo, branch, file_extensions = "")
}
\arguments{
\item{directory}{A character string giving the full name of the Github
folder with the theme to be downloaded.}

\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}

\item{branch}{A character string giving the branch in the repository}

\item{file_extensions}{A character vector with file extensions (excl. '.',
e.g. c("R", "Rmd"))}
}
\description{
The function downloads all the files from the Github repository into a
folder with the same name as the subfolder in the online repository
}
