% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_info}
\alias{github_api_info}
\title{Get Github API information}
\usage{
github_api_info(user, repo)
}
\arguments{
\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\value{
A vector with the list of API information
}
\description{
The function gets the API information for a given user and Github repository.
}
