% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_default_branch}
\alias{github_default_branch}
\title{Get the default branch of a repository}
\usage{
github_default_branch(user, repo)
}
\arguments{
\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\value{
A character string with the default branch
}
\description{
Get the default branch of a repository
}
