% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_download_files}
\alias{github_download_files}
\title{Download Github Director}
\usage{
github_download_files(file_list, user, repo, branch)
}
\arguments{
\item{file_list}{A vector of files}

\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}

\item{branch}{A character string giving the branch in the repository}
}
\description{
Download Github Director
}
