% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_repo_file_list}
\alias{github_repo_file_list}
\title{Get the list of files in the Github repository}
\usage{
github_repo_file_list(user, repo)
}
\arguments{
\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\value{
A vector with the list of files contained in the repository
}
\description{
Get the list of files in the Github repository
}
