% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo.R
\docType{package}
\name{gizmo}
\alias{gizmo}
\alias{gizmo-package}
\title{gizmo: A Collection of Utility Functions for Choice Modeling and Statistical Analysis}
\description{
All the useful utility functions and tests in one convenient place. Gizmo 
contains a set of utilities and tests that I use frequently in my own work.
}
