% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize-new-project.R
\name{initialize_new_project}
\alias{initialize_new_project}
\title{Initialize a new project}
\usage{
initialize_new_project(theme = "nmbu", overwrite = FALSE)
}
\arguments{
\item{theme}{A character string giving the name of the Beamer theme}

\item{overwrite}{A boolean equal to TRUE if existing folders and files
should be overwritten. All warnings are supressed. Default is FALSE.}
}
\description{
Initialize a new project
}
