% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_common_packages.R
\name{install_common_packages}
\alias{install_common_packages}
\title{Wrapper function to install commonly used packages}
\usage{
install_common_packages()
}
\description{
The function should only be called with a fresh install of R. It will install
all packages I have added to the list. The function serves no other purpose.
}
