% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{is_package_loaded}
\alias{is_package_loaded}
\title{Check whether a package is loaded}
\usage{
is_package_loaded(pkgs)
}
\arguments{
\item{pkgs}{A string of length >= 1 of package names.}
}
\value{
TRUE if all packages are loaded and FALSE if one or 
more of the packages are not loaded.
}
\description{
A simple check to see whether a package is loaded.
}
