% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last}
\alias{last}
\title{Get the last element of a vector}
\usage{
last(x)
}
\arguments{
\item{x}{A vector, matrix or list}
}
\value{
A single value or list element
}
\description{
\code{last} extracts the last element of a vector, matrix or list
}
\examples{
x <- 1:4
last(x)

x <- c("my", "name", "is", "buttons")
last(x)

x <- matrix(runif(10), 5, 2)
last(x)

x <- matrix(LETTERS[1:10], 5, 2)
last(x)

x <- list(1:5, 6:10)
last(x)

}
