% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{load_packages}
\alias{load_packages}
\title{Load multiple packages}
\usage{
load_packages(pkgs, silent = FALSE)
}
\arguments{
\item{pkgs}{A string of length >= 1 of package names.}

\item{silent}{A boolean. If TRUE suppresses package loading messages. Default
FALSE.}
}
\description{
A wrapper function to quickly load multiple packages passed as a 
character string.
}
