% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_file_extensions_regex}
\alias{make_file_extensions_regex}
\title{Create a regex for file extenstions}
\usage{
make_file_extensions_regex(string)
}
\arguments{
\item{string}{A character vector with file extensions}
}
\description{
Create a regex for file extenstions
}
\examples{
\dontrun{
  make_file_extensions_regex(c("jpg"))
  make_file_extensions_regex(c("jpg", "eps", "bmp"))
}
}
