% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Arrange multiple plots in a grid}
\usage{
multiplot(..., plot_list = NULL, columns = 1, by_row = FALSE,
  layout_matrix = NULL)
}
\arguments{
\item{...}{plot objects}

\item{plot_list}{List of plot objects. Defaults to NULL.}

\item{columns}{An integer specifying the number of columns. The number of rows
is calculated automatically. Defaults to 1.}

\item{by_row}{If TRUE, the generated layout_matrix is filled by row. Default
is FALSE.}

\item{layout_matrix}{A layout matrix. Defaults to NULL.}
}
\description{
The function takes a list of plots made using ggplot() as input and arranges 
them in a grid.
}
\examples{
library(ggplot2)
library(tibble)

df <- tibble(x = seq_len(10), y = runif(10), z = runif(10) * 1.5)
plot_1 <- ggplot(data = df) + geom_point(aes(x = x, y = y))
plot_2 <- ggplot(data = df) + geom_point(aes(x = x, y = z))

multiplot(plot_1, plot_2)
multiplot(plot_1, plot_2, columns = 2)
multiplot(plot_list = list(plot_1, plot_2))

}
\references{
The function is based on the multiplot function in "Cookbook for R". Please
see the original code at 
<http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/>
}
