% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe_test.R
\name{poe_test}
\alias{poe_test}
\title{Test for the difference in empirical distributions}
\usage{
poe_test(x, y)
}
\arguments{
\item{x}{An input vector}

\item{y}{An input vector}
}
\value{
A numeric test value
}
\description{
The function calculates the complete combinatorial of the supplied vectors
using a loop implementation by taking the difference between every 
combination of the elements in the vectors. The input vectors must 
be numeric, but can be of different lengths.
}
\examples{
x <- qnorm(runif(100), mean = -0.5, sd = 1)
y <- qnorm(runif(100), mean = 1.5, sd = 2)
poe_test(x, y)

}
\references{
Poe G. L., Giraud, K. L. & Loomis, J. B., 2005, Computational methods for 
measuring the difference of empirical distributions, American Journal of
 Agricultural Economics, 87:353-365
}
