% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_dir}
\alias{rename_dir}
\title{Rename a directory}
\usage{
rename_dir(old, new)
}
\arguments{
\item{old}{String with old name (path)}

\item{new}{String with new name (path)}
}
\description{
Rename a directory
}
