% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_folder}
\alias{rename_folder}
\title{Rename a folder}
\usage{
rename_folder(old, new)
}
\arguments{
\item{old}{String with old name (path)}

\item{new}{String with new name (path)}
}
\description{
Rename a folder
}
