% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{setup_models}
\alias{setup_models}
\title{Setup models folder}
\usage{
setup_models(directory = "examples", user = "edsandorf", repo = "cmdlR")
}
\arguments{
\item{directory}{A character string giving the full name of the Github
folder with the theme to be downloaded.}

\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\description{
A wrapper around \code{\link{download_from_github}} to make it easy to 
download and setup a 'models' folder within the current project using
examples from Github. The function downloads the specified examples and
renames it to 'models'.
}
