% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe_test.R
\name{summary_poe_test}
\alias{summary_poe_test}
\title{Summary function for poe_test()}
\usage{
summary_poe_test(x)
}
\arguments{
\item{x}{An estimate obtained from \code{\link{poe_test}}.}
}
\description{
The function prints a more descriptive output to the console with the result
of the test for difference in empirical distributions.
}
\examples{
x <- qnorm(runif(100), mean = -0.5, sd = 1)
y <- qnorm(runif(100), mean = 1.5, sd = 2)
test_stat <- poe_test(x, y)
summary_poe_test(test_stat)

}
