% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{update_models}
\alias{update_models}
\title{Update the examples in the models folder}
\usage{
update_models(directory = "examples", user = "edsandorf", repo = "cmdlr")
}
\arguments{
\item{directory}{A character string giving the full name of the Github
folder with the theme to be downloaded.}

\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\description{
A wrapper around \code{\link{update_from_github}} to make it easy to update
model examples in the current project with new and updated examples from
Github.
}
