% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizmo.R
\docType{package}
\name{gizmo}
\alias{gizmo}
\alias{gizmo-package}
\title{gizmo: A supplementary package for statistical analysis and choice modeling}
\description{
The gizmo package provides several useful functions for statistical analysis
and choice modeling.
}
