% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{setup_beamer}
\alias{setup_beamer}
\title{Setup Beamer}
\usage{
setup_beamer(theme = "nmbu", user = "edsandorf", repo = "latex-templates")
}
\arguments{
\item{theme}{A character string giving the name of the Beamer theme}

\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\description{
Downloads the correct Beamer template folder from my Github repository and
renames it to 'presentation'
}
