% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{update_beamer}
\alias{update_beamer}
\title{Updates the Beamer template}
\usage{
update_beamer(theme = "nmbu", user = "edsandorf", repo = "latex-templates")
}
\arguments{
\item{theme}{A character string giving the name of the Beamer theme}

\item{user}{A character string giving the user}

\item{repo}{A character string giving the repository}
}
\description{
The function updates the Beamer templates in the current working directory
and all sub-directories recursively.
}
